feat(deps): bump rainlang to 0.1.8 for NotAnAddress(bytes32)
rainlang 0.1.8 ships the rainlang#539 ruling that NotAnAddress carries the
raw bytes32 stack word rather than a uint256. The two vault-word guards now
pass the word directly (revert NotAnAddress(vault)), the expectRevert
encodings pass bytes32, and a selector pin test locks the bytes32 ABI form
against the signature string. rain-deploy 0.1.3 and rain-extrospection 0.1.1
come along as rainlang 0.1.8's pinned transitives (recursive_deps = false).
Pointers regenerated: the words bytecode embeds the new selector, so
BYTECODE_HASH and function pointers moved.

Closes #245

+++ src/lib/op/erc4626/LibOpERC4626ConvertToShares.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.25;
 import {OperandV2, StackItem} from "rain-interpreter-interface-0.1.0/src/interface/IInterpreterV4.sol";
 import {Float} from "rain-math-float-0.1.1/src/lib/LibDecimalFloat.sol";
 import {LibERC4626, IERC4626Minimal} from "../../erc4626/LibERC4626.sol";
-import {NotAnAddress} from "rainlang-0.1.2/src/error/ErrRainType.sol";
+import {NotAnAddress} from "rainlang-0.1.8/src/error/ErrRainType.sol";
 
 library LibOpERC4626ConvertToShares {
     /// Extern integrity for erc4626-convert-to-shares.
@@ -40,7 +40,7 @@ library LibOpERC4626ConvertToShares {
         // of vault as an address.
         // Casting to `uint160` is intentional to detect non-address values.
         //forge-lint: disable-next-line(unsafe-typecast)
-        if (uint256(vault) != uint256(uint160(uint256(vault)))) revert NotAnAddress(uint256(vault));
+        if (uint256(vault) != uint256(uint160(uint256(vault)))) revert NotAnAddress(vault);
 
         // Casting to `uint160` is safe because `NotAnAddress` above
         // ensures the value fits in 160 bits.
